Refactor ArticlePage to use ArticleContent component
This refactor moves article fetching logic into a separate ArticleContent component, improving code modularity and readability. It removes the redundant fetching logic in ArticlePage, leveraging the new ArticleContent for a cleaner implementation.

Took 4 minutes
